+++ src/dotnetCampus.Ipc/IpcRouteds/DirectRouteds/IpcDirectRoutedMessageWriter.cs
@@ -1,0 +1,14 @@
+﻿#if NET6_0_OR_GREATER
+using System.IO;
+
+namespace dotnetCampus.Ipc.IpcRouteds.DirectRouteds;
+
+static class IpcDirectRoutedMessageWriter
+{
+    public static void WriteHeader(BinaryWriter writer, ulong businessMessageHeader, string routedPath)
+    {
+        writer.Write(businessMessageHeader);
+        writer.Write(routedPath);
+    }
+}
+#endif
