+++ src/dotnetCampus.Ipc/IpcRouteds/DirectRouteds/JsonIpcDirectRoutedContext.cs
@@ -1,0 +1,5 @@
+﻿#if NET6_0_OR_GREATER
+namespace dotnetCampus.Ipc.IpcRouteds.DirectRouteds;
+
+public record JsonIpcDirectRoutedContext(string PeerName);
+#endif
